+++ api/controllers/discipline/remove.js
@@ -1,3 +1,5 @@
+/* global sails, Discipline */
+
 module.exports = {