+++ api/helpers/discipline/add.js
@@ -2,15 +2,22 @@
 
 module.exports = {
 
-    friendlyName: 'discipline.banDay',
+    friendlyName: 'discipline.add',
 
-    description: 'Ban a specified host for 24 hours.',
+    description: 'Add a disciplinary action to a specified host.',
 
     inputs: {
         host: {
             required: true,
             type: 'string',
             description: 'The unique ID assigned to the host that we are banning.'
+        },
+        
+        action: {
+            type: "string",
+            required: true,
+            isIn: ["dayban", "permaban", "showban"],
+            description: "Type of ban: dayban (24 hours from createdAt), permaban (indefinite), show ban (until the current broadcast ends)."
         },
         
         message: {
@@ -35,7 +42,7 @@ module.exports = {
             await sails.helpers.messages.removeMass(inputs.host);
 
             // Add a dayban to the database
-            var reference = await Discipline.create({active: inputs.active, IP: inputs.host, action: 'dayban', message: inputs.message}).fetch();
+            var reference = await Discipline.create({active: inputs.active, IP: inputs.host, action: inputs.action, message: inputs.message}).fetch();
 
             // Broadcast the ban to the client
             sails.sockets.broadcast(`discipline-${inputs.host.replace('website-', '')}`, `discipline`, {"discipline": `Your interactions with WWSU have been placed under review. Please email engineer@wwsu1069.org for further assistance. Please include the following reference number(s) in your email: ${reference.ID}`});